workitt-backend/requirements.txt
Flask
Flask-Login
Flask-Cors
cryptography
flask-sqlalchemy
bcrypt
openai
pyyaml
marshmallow
pdfplumber
flask-session